Rename package and main class
