+++ src/main/java/com/github/pocketkid2/moongenerator/FlagPopulator.java
@@ -1,4 +1,4 @@
-package com.dinnerbone.bukkit.moon;
+package com.github.pocketkid2.moongenerator;
 
 import java.util.Random;
